app/requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
google-generativeai==0.5.4
pydantic==2.11.9
sendgrid==6.12.3
python-http-client==3.3.7
motor==3.7.1
pymongo==4.13.0
dnspython==2.7.0
redis==5.0.1
requests==2.32.3
aiofiles==0.8.0
pillow==11.2.1
timezonefinder==6.5.9
Jinja2==3.1.6
python-multipart==0.0.20
pytz==2023.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
pyotp==2.9.0
